refactor(fsd): move RoleContext+useRole to entities/user/model (W3)
Features now import role state from @entities layer instead of
cross-importing from @features/auth, fixing FSD slice isolation.
