+++ frontend/src/components/layout/Sidebar.tsx
@@ -10,7 +10,7 @@ import {
   Code2,
 } from 'lucide-react';
 import { cn } from '@shared/lib/cn';
-import { useRole } from '@features/auth/model/useRole';
+import { useRole } from '@entities/user/model/useRole';
 
 interface NavItem {
   to: string;
