+++ frontend/src/entities/user/index.ts
@@ -1,3 +1,5 @@
 export { getMe } from './api/userApi';
 export type { AuthUser, AuthRole } from './api/userApi';
 export { userQueryKeys } from './api/userQueryKeys';
+export { RoleProvider, RoleContext } from './model/RoleContext';
+export { useRole } from './model/useRole';
