+++ frontend/src/features/auth/index.ts
@@ -1,6 +1,6 @@
 export { AuthProvider, AuthContext } from './model/AuthContext';
 export type { AuthContextValue, AuthUser } from './model/AuthContext';
 export { useAuth } from './model/useAuth';
-export { RoleProvider, RoleContext } from './model/RoleContext';
-export { useRole } from './model/useRole';
+export { RoleProvider, RoleContext } from '@entities/user/model/RoleContext';
+export { useRole } from '@entities/user/model/useRole';
 export { LoginForm } from './ui/LoginForm';
